requirements.txt
discord.py>=2.4.0,<3.0.0
google-genai>=0.8.0,<1.0.0
python-dotenv>=1.0.1,<2.0.0
pydantic>=2.8.0,<3.0.0
openpyxl>=3.1.0,<4.0.0
gspread>=6.0.0,<7.0.0
pytest>=8.3.0,<9.0.0
pytest-asyncio>=0.24.0,<1.0.0